pyproject.toml
[tool.poetry]
name = "django-mini-project"
version = "0.1.0"
description = "A minimal Django project setup"
authors = ["Your Name <you@example.com>"]
package-mode = false  # 패키지 모드 비활성화

[tool.poetry.dependencies]
python = "^3.9"
django = "^3.2"
djangorestframework = "^3.12"
drf-yasg = "^1.21.9"
psycopg2-binary = "^2.9.10"
django-environ = "^0.12.0"

[tool.poetry.group.dev.dependencies]
drf-yasg = "^1.20"
django-stubs = "^5.1.3"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
types-psycopg2 = "^2.9.21.20250121"


[tool.poetry.group.black.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = "3.12"
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.coverage.report]
fail_under = 90

[tool.coverage.run]
omit = [
    ".mypy_cache/**/*",
    "*/migrations/*",
    "*/*test*.py",
]